Fix poweroff command and ON/OFF button in imx_system_off()
When using poweroff command in Linux or pressing the ON/OFF button, the
system correctly shuts down. But instead of staying off, it immediately
starts again.

In a previous commit, NXP has unconditionally activated the alarm setting
of the real time counter when switching off the system power. This assumes
that the rtc-snvs driver is active in Linux, that correctly sets the alarm
time and resets the alarm event. There is always an alarm event on a cold
start because both the real time counter and the alarm time are both 0 and
therefore match, causing the alarm event to be triggered.

On F&S systems, there is an external RTC and the internal rtc-snvs is not
used. This means that the SNVS alarm time is not set and the alarm event
from power-on is still latched. When shutting down, this event immediately
triggers a reboot because of the alleged alarm. And as this alarm flag is
never cleared, the behaviour also doesn't change on subsequent shutdowns.

In fact, setting the alarm bits in SNVS_LPCR register is not necessary when
shutting down because these bits are already set by the rtc-snvs driver.
Therefore in imx_system_off(), instead of writing the whole register
contents, just add the power-off bits. This should work in both cases,
with and without rtc-snvs driver.