plat:imx:imx93: allow OCRAM access in NonSecure state
The BOARD-CFG is placed in OCRAM. To allow access during U-Boot, the MBC
is configured to allow access to the FDT-BLOB.

This is a temporarly change!
It is intended to configure the MBC so that the necessary 4KiB blocks
are enabled. The entire OCRAM is currently being enabled for non-secure
access.

The necessary configurations of the first two blocks for the 8KiB of the
BOARD-CFG must be determined.