imx:imx8ulp:psci: ensures TPM5 is enabled during suspend
During testing of DPD and PD modes, it was found that the WakeUp
function did not work with FreeRTOS when PD was used.

FreeRTOS calls the UPOWER API (upwr_xcp_set_rtd_apd_llwu), which is
supposed to wake up the APD.

However, the Cortex-A could never be woken up by WUU1 (triggered by
UPOWER). For some unknown reason, this issue is related to TPM5.

Workaround: ensure that TPM5 is always active during suspend.