plat:imx8ulp: check activ clock before register dump
To ensure that the ATF always behaves stably on the imx8ulp platform
when switching between suspend and wake-up contexts, it must always be
ensured that the necessary prerequisites, such as an active clock, are
met during the register dump.

In app_context.c, two issues occur in this regard.

Issue 1: It is blindly assumed that the DBG-UART is LPUART5:
This is not always the case, nor is there a fixed rule that LPUART5 must
be used as DBG. If registers need to be dumped and restored for a
SUSPEND/Wakeup scenario, it must always be checked beforehand whether
the clock is even on. The kernel itself controls the clock configuration
and can also be switched off via userspace interactions. In this case,
the clock for LPUART, including for the DBG port, must not be active.

FIX: Save base-addr of dbg uart during boot and check if clock is
enabled during suspend/wakeup.

Issue 2: It is blindly assumed that tpm5/6 is used as timer-tpm:
As with the UART, the TPM does not check whether the corresponding clock
is active.

FIX: Check CGC register in the respective PCC. Remove tpm5 and expect
TPM6 as the default timer.